Two column layout
Previous commits, the code from the book, simply ran together the label
and the assignee. This commit lays out these items in a two column
layout.

To achieve this layout, because assignee was sometimes empty, I needed
to insert a non-breaking space; otherwise, the browser added the next
`div` element to the row. This requirement led to a function which
return either the `assignee` or the Unicode non-breaking space
character, `\u00a0`.