Chapter 2 ng-repeat over an object
Although the book indicates that Angular will sort the keys in
alphabetical order, Angular 1.4 introduced a breaking change (see
http://brianhann.com/angular-1-4-breaking-changes-to-be-aware-of/#ngrepeat).

Because of this breaking change the keys are now returned in a
browser-dependent order. If I wish to see the items in a specific order,
I must sort them myself. The referenced article refers to another
article on how to sort them.