Visually distinguish closely related "rows"
Previous commits presented all elements using the same style. As a
result, one had to "know" that the label, "First Note," and its
corresponding done flag (true), were related.

This commit used the `ng-class` directive to invoke a function to
determine the class of the element. Elements with the 'odd' class have a
green background; elements with the 'even' class have a yellow
background. The class is determined by the function `isEvenOdd`, defined
on the controller.