First example of `ng-model`
It works - but it does not display the text "nothing" in the input box
when it first renders the page. I'm uncertain the cause for this
"failed" expectation. It could be a newer version of AngularJS (with a
change of behavior) or it could be I have made a small mistake in the
HTML / script.